Move location, rename, and make 'exportTemplate' file the main simulation
Now we do not have to update two separate files when adding in a new dashboard input
