+++ src/python/impactx/dashboard/Toolbar/controls.py
@@ -13,7 +13,7 @@ from ..Analyze.plotsMain import available_plot_options, load_dataTable_data, upd
 from ..Input.components.card import CardComponents
 from ..Input.generalFunctions import generalFunctions
 from ..Run.controls import run_execute_impactx_sim
-from .exportTemplate import input_file
+from ..Run.simulation import input_file
 from .importParser import DashboardParser
 
 server, state, ctrl = setup_server()
